chore: reconcile CLI tests and sync metadata
