+++ .pdd/meta/metadata_sync_python.json
@@ -1,8 +1,8 @@
 {
   "pdd_version": "0.0.305.dev0",
-  "timestamp": "2026-07-18T01:17:36+00:00",
+  "timestamp": "2026-07-18T06:49:47.463528+00:00",
   "command": "fix",
-  "prompt_hash": "45946e86e4b0ce8347b0e03c2c7655ec011a8103df6597a89cc7d445d0584062",
+  "prompt_hash": "2a71c89e91bbffa0e7bc56f96abd1505d8d045c6899aeeea1998a60266439990",
   "code_hash": "3afaa5a409e500b3f91aab33dacee33fae47447fba80a1702d65a33761261782",
   "example_hash": "704453b3386724eef225b61c48f2fe922e1a973ac211c494e1d955dd44bcca11",
   "test_hash": "82af9e0bff7550d1f9bf46f4c6e7f4792143565e4e228d1507976f6b9fbac5f6",
@@ -13,7 +13,7 @@
     "context/python_preamble.prompt": "0388ed131bf986f8752e1bc4c81e4da0460cfe2908ec8c60b1314edbab768254",
     "pdd/architecture_sync.py": "0e0ff74978f49d629f92365723ebbea0e7c70e6d5234ccfdafb1c8e09c110c11",
     "pdd/architecture_registry.py": "5722225846b9cc8189b410a78e5714b4f2916d5748271d27f6f738692429c309",
-    "pdd/sync_determine_operation.py": "2e6e0d9f643ed7244e3dddb5c9a3337692862ff6cbfafe338ab0af2ccbdc06e3",
+    "pdd/sync_determine_operation.py": "38b0ceb109e16b9a42cae5df6e97299e9983d808c25e7f9d99f3791976033c17",
     "pdd/operation_log.py": "081b3e6e93ef7090d29ba4362edae46772a371e16df5d1d962814ad3a8069b60"
   }
 }
