+++ .pdd/meta/sync_determine_operation_python.json
@@ -1,9 +1,9 @@
 {
   "pdd_version": "0.0.305.dev0",
-  "timestamp": "2026-07-18T01:17:36+00:00",
+  "timestamp": "2026-07-18T06:49:47.057431+00:00",
   "command": "fix",
   "prompt_hash": "d78eee431bf5772c617f4abb4422a245f123f6efd52c0b0924074af8b9a22c3c",
-  "code_hash": "2e6e0d9f643ed7244e3dddb5c9a3337692862ff6cbfafe338ab0af2ccbdc06e3",
+  "code_hash": "38b0ceb109e16b9a42cae5df6e97299e9983d808c25e7f9d99f3791976033c17",
   "example_hash": "225d20ea0210bbf52650a9f424879ff495e574dd2411eb4b86592a6d06cf59a5",
   "test_hash": "47af96d6510d3e8a8416f8f1e219d6c228db5db23e106101af03388f391577b2",
   "test_files": {
